propagating naming through SOM summarizers
