+++ examples/estimation_examples/README.md
@@ -4,6 +4,6 @@ This directory contains example notebooks explaining how to use the RAIL Estimat
 
 - **RAIL_estimation_demo.ipynb** explains to inform a model used for photo-z estimation, and then how to use that model to estimate p(z).
 
-- **SimpleSOM_demo.ipynb**, **somocluSOM_demo.ipynb**, and **somocluSOMcluster_demo.ipynb** demonstrate of the use of the `SimpleSOMSummarizer` summarization module.
+- **SimpleSOM_demo.ipynb**, **somocluSOM_demo.ipynb**, and **somocluSOMcluster_demo.ipynb** demonstrate of the use of the `SummarizeSimpleSOM` and `SummarizeSOMoclu` `SZPZSummarizer` stages.
 
 - [test_sampled_summarizers.ipynb](https://lsstdescrail.readthedocs.io/en/latest/source/estimation-notebooks.html#testing-sampled-summarizers) outlines quick and dirty bootstrap versions of the `NaiveStack`, `PointEstimateHist`, and `VarInference` sumarizers, as well as `NZDir`.
