+++ src/rail/estimation/algos/somocluSOM.py
@@ -13,28 +13,30 @@ from mpl_toolkits.axes_grid1 import make_axes_locatable
 import sklearn.cluster as sc
 from scipy.spatial.distance import cdist
 from rail.core.common_params import SHARED_PARAMS
+from rail.core.algo_utils import _computemagcolordata
+
+# TODO: review whether these pragmas are justified
 
 
-
-def _computemagcolordata(data, ref_column_name, column_names, colusage):
-    if colusage not in ['colors', 'magandcolors', 'columns']:  # pragma: no cover
-        raise ValueError(f"column usage value {colusage} is not valid, valid values are 'colors', 'magandcolors', and 'columns'")
-    numcols = len(column_names)
-    if colusage == 'magandcolors':
-        coldata = np.array(data[ref_column_name])
-        for i in range(numcols - 1):
-            tmpcolor = data[column_names[i]] - data[column_names[i + 1]]
-            coldata = np.vstack((coldata, tmpcolor))
-    if colusage == 'colors':
-        coldata = np.array(data[column_names[0]] - data[column_names[1]])
-        for i in range(numcols - 2):
-            tmpcolor = data[column_names[i + 1]] - data[column_names[i + 2]]
-            coldata = np.vstack((coldata, tmpcolor))
-    if colusage == 'columns':  # pragma: no cover
-        coldata = np.array(data[column_names[0]])
-        for i in range(numcols - 1):
-            coldata = np.vstack((coldata, np.array(data[column_names[i + 1]])))
-    return coldata.T
+# def _computemagcolordata(data, ref_column_name, column_names, colusage):
+#     if colusage not in ['colors', 'magandcolors', 'columns']:  # pragma: no cover
+#         raise ValueError(f"column usage value {colusage} is not valid, valid values are 'colors', 'magandcolors', and 'columns'")
+#     numcols = len(column_names)
+#     if colusage == 'magandcolors':
+#         coldata = np.array(data[ref_column_name])
+#         for i in range(numcols - 1):
+#             tmpcolor = data[column_names[i]] - data[column_names[i + 1]]
+#             coldata = np.vstack((coldata, tmpcolor))
+#     if colusage == 'colors':
+#         coldata = np.array(data[column_names[0]] - data[column_names[1]])
+#         for i in range(numcols - 2):
+#             tmpcolor = data[column_names[i + 1]] - data[column_names[i + 2]]
+#             coldata = np.vstack((coldata, tmpcolor))
+#     if colusage == 'columns':  # pragma: no cover
+#         coldata = np.array(data[column_names[0]])
+#         for i in range(numcols - 1):
+#             coldata = np.vstack((coldata, np.array(data[column_names[i + 1]])))
+#     return coldata.T
 
 
 def get_bmus(som, data=None, split=200):  # pragma: no cover
@@ -123,7 +125,7 @@ def plot_som(ax, som_map, grid_type='rectangular', colormap=cm.viridis, cbar_nam
     ax.axis('off')
 
 
-class Inform_somocluSOMSummarizer(CatInformer):
+class SOMocluInformer(CatInformer):
     """Summarizer that uses a SOM to construct a weighted sum
     of spec-z objects in the same SOM cell as each photometric
     galaxy in order to estimate the overall N(z).  This is
@@ -169,7 +171,7 @@ class Inform_somocluSOMSummarizer(CatInformer):
       pickle file containing the `somoclu` SOM object that
     will be used by the estimation/summarization stage
     """
-    name = 'Inform_SOMoclu'
+    name = 'SOMocluInformer'
     config_options = CatInformer.config_options.copy()
     config_options.update(nondetect_val=SHARED_PARAMS,
                           mag_limits=SHARED_PARAMS,
@@ -233,7 +235,7 @@ class Inform_somocluSOMSummarizer(CatInformer):
         self.add_data('model', self.model)
 
 
-class somocluSOMSummarizer(SZPZSummarizer):
+class SOMocluSummarizer(SZPZSummarizer):
     """Quick implementation of a SOM-based summarizer. It will
     group a pre-trained SOM into hierarchical clusters and assign
     a galaxy sample into SOM cells and clusters. Then it
@@ -298,7 +300,7 @@ class somocluSOMSummarizer(SZPZSummarizer):
     qp_ens: qp Ensemble
       ensemble of bootstrap realizations of the estimated N(z) for the input photometric data
     """
-    name = 'somocluSOMSummarizer'
+    name = 'SOMocluSummarizer'
     config_options = SZPZSummarizer.config_options.copy()
     config_options.update(zmin=SHARED_PARAMS,
                           zmax=SHARED_PARAMS,
@@ -544,3 +546,4 @@ class somocluSOMSummarizer(SZPZSummarizer):
         self._cellid_handle.set_data(id_dict, partial=True)
         self._cellid_handle.write_chunk(start, end)
 
+# TODO: add EstimateSOMoclu
