+++ tests/estimation/test_som_summarizers.py
@@ -61,8 +61,8 @@ def one_algo(key, inform_class, summarizer_class, summary_kwargs):
 
 def test_SimpleSOM():
     summary_config_dict = {"m_dim": 21, "n_dim": 21, "column_usage": "colors"}
-    inform_class = simpleSOM.InformSimpleSOM
-    summarizerclass = simpleSOM.SummarizeSimpleSOM
+    inform_class = simpleSOM.SimpleSOMInformer
+    summarizerclass = simpleSOM.SimpleSOMSummarizer
     _ = one_algo("SimpleSOM", inform_class, summarizerclass, summary_config_dict)
 
 
@@ -73,6 +73,6 @@ def test_SimpeSOM_with_mag_and_colors():
         "column_usage": "magandcolors",
         "objid_name": "id",
     }
-    inform_class = simpleSOM.InformSimpleSOM
-    summarizerclass = simpleSOM.SummarizeSimpleSOM
+    inform_class = simpleSOM.SimpleSOMInformer
+    summarizerclass = simpleSOM.SimpleSOMSummarizer
     _ = one_algo("SimpleSOM_wmag", inform_class, summarizerclass, summary_config_dict)
